Split bundle processor into separate class.
